Move get_keypair to a shared location
